api/requirements.txt
asgiref==3.8.1
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
ghp-import==2.1.0
idna==3.10
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.44
mkdocs-material-extensions==1.3.1
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
Pygments==2.18.0
pymdown-extensions==10.12
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
